+++ apps/maimai-web/src/lib/components/profile/Rating.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+  import { getMaimaiRatingLevel } from "@repo/core/maimai";
   import { cn } from "@repo/ui/utils";
 
   interface Props {
@@ -8,31 +9,7 @@
 
   let { rating, calculatedRating }: Props = $props();
 
-  let ratingLevel = $derived(
-    rating >= 16000
-      ? "rainbow_kiwami"
-      : rating >= 15000
-        ? "rainbow"
-        : rating >= 14500
-          ? "platinum"
-          : rating >= 14000
-            ? "gold"
-            : rating >= 13000
-              ? "silver"
-              : rating >= 12000
-                ? "bronze"
-                : rating >= 10000
-                  ? "purple"
-                  : rating >= 7000
-                    ? "red"
-                    : rating >= 4000
-                      ? "orange"
-                      : rating >= 2000
-                        ? "green"
-                        : rating >= 1000
-                          ? "blue"
-                          : "normal",
-  );
+  let ratingLevel = $derived(getMaimaiRatingLevel(rating));
 
   let ratingMatched = $derived(rating === calculatedRating);